Final HUD using sprites from the game. We now have age_panel displaying the current player age, the resources panel, the action_panel, the selection panel and the minimap panel (for now only displayed when examine entity for performances issues).
I began to write a action_panel class and associated method maybe for later use (each entity may have a different action menu to display after all). The tooltips and all the hud should have been updated to be now screensize adaptable, and costs are displayed in red when not enough resources.