added villager animation (the attack one is weird, i can't fixe it. I'd need to use anchor points, but i can't manage to make it work.
created a "get angle between 2 tiles" method to adapt the angle we display the village sprite.
Fixed a bug when displaying sprites (always turning left before moving). Was caused by a difference between pos (,) and [,]

fixed bug when loading images from folder for animation (not displayed in the correct order). had to understand how the system orders it.