Demo build sprite of villager
Showing a demo "build" of a villager. I have built lots of function to achieve this so i dont want to waste. If it is good, i can extract all the others action. Use pygame.transform.flip to flip the image, so i dont extract the others hafl of sprite